requirements.txt
asgiref==3.8.1
boto3==1.34.139
botocore==1.34.139
django-storages==1.14.3
Django==5.0.6
django-environ==0.11.2
djangorestframework==3.15.2
drf-yasg==1.21.7
google==3.0.0
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.2
gunicorn==22.0.0
inflection==0.5.1
mysql==0.0.3
mysqlclient==2.2.4
openai==1.35.10
openapi==2.0.0
packaging==24.1
PyMySQL==1.1.1
pytz==2024.1
PyYAML==6.0.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
logger~=1.4
numpy==2.0.0
opencv-python-headless==4.10.0.84
moviepy==1.0.3
requests==2.32.3
celery==5.4.0
django-celery-beat==2.6.0
redis==5.0.7
flower==2.0.1
elasticsearch-dsl>=7.0.0,<8.0.0
django-elasticsearch-dsl>=7.0.0,<8.0.0
django-cors-headers==4.4.0